Allow passing flags in `ga`
This is a fairly naive implementation. In particular it will fall over
if you need to remove a file that begins with a '-'. A better
implementation would ask git if the file exists in the index, and draw
conclusions about file vs. flag based on that, rather than just assuming
anything that starts with a hyphen is a flag.